% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boaR-package.R
\docType{package}
\name{boaR-package}
\alias{boaR}
\alias{boaR-package}
\title{boaR: Predict Wild Pig Abundance Using A Removal Model}
\description{
Predicts wild pig abundance using a removal model.
}
\author{
\strong{Maintainer}: John Foster \email{john.foster@usda.gov}

}
\keyword{internal}
