% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble_removal_model.R
\name{calc_log_potential_area}
\alias{calc_log_potential_area}
\title{Calculate the log of the potential area surveyed}
\usage{
calc_log_potential_area(
  log_rho,
  log_gamma,
  p_unique,
  log_effort_per,
  effort_per,
  n_trap_m1,
  log_pi,
  method
)
}
\arguments{
\item{log_rho}{vector of log rho coefficients (scaling effort to area) for each method}

\item{log_gamma}{vector of log gamma coefficients (saturating effect) for traps and snares}

\item{p_unique}{vector of probabilities that a trap captures a unique individual for traps and snares}

\item{log_effort_per}{scalar of log effort per survey}

\item{effort_per}{scalar of effort per survey}

\item{n_trap_m1}{scalar of number of traps minus 1 for each survey}

\item{log_pi}{log of pi}

\item{method}{method index for each survey}
}
\value{
log of the potential area surveyed
}
\description{
Calculate the log of the potential area surveyed
}
