% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt3.R
\name{filt3}
\alias{filt3}
\title{filt3}
\usage{
filt3(lon, lat, ingrid, grid5 = grid5(lon, lat, ingrid))
}
\arguments{
\item{lon}{longitude (vector) of the satellite image}

\item{lat}{latitude (vector) of the satellite image}

\item{ingrid}{The satellite data (matrix)}

\item{grid5}{results of \code{\link{filt5}}}
}
\value{
returns a median smoothed grid of satellite data
}
\description{
filt3
}
\examples{
# none here
}
\seealso{
\code{\link{boa}}
}
