% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_flags.R
\name{get_model_flags}
\alias{get_model_flags}
\title{Create boolean values to build the correct model}
\usage{
get_model_flags(df)
}
\arguments{
\item{constants}{The constants list the is used to build a nimble model.}
}
\value{
a list of values
single_property is TRUE of only one property is in the data.
single_method is TRUE of only one removal method is used.
use_shooting is TRUE if firearms, fixed-wing aircraft, or helicopters are used.
use_traps_and_snares is TRUE if both traps and snares are used.
use_traps_or_snares is TRUE if one of traps or snares are used, but not both.
}
\description{
Create boolean values to build the correct model
}
