% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_parallel.R
\name{mcmc_parallel}
\alias{mcmc_parallel}
\title{Function for running nimble in parallel}
\usage{
mcmc_parallel(
  n_chains,
  model_constants,
  model_data,
  model_flags,
  params_check,
  n_iters,
  dest,
  monitors_add = NULL,
  custom_samplers = NULL,
  export = NULL,
  ...
)
}
\arguments{
\item{n_chains}{The number of chains to run in parallel. Used to make a socket cluster with n_chains nodes.}

\item{model_constants}{model constants, as a list}

\item{model_data}{model data, as a list}

\item{params_check}{vector of parameters (nodes) to assess convergence}

\item{n_iters}{the number of iterations to run in each chunk after the model has compiled}

\item{dest}{output filepath. within this directory each 'chunk' will be saved into it's own directory}

\item{monitors_add}{vector of nodes to monitor in addition to default nodes}

\item{custom_samplers}{data frame of mcmc config. Nodes in one column, sampler types in another}

\item{model_inits}{initial model values, as a list, one list element for each chain}
}
\description{
Run nimble on specified cluster, check for convergence and save samples periodically
}
