% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble_constants.R
\name{nimble_constants}
\alias{nimble_constants}
\title{Create A List For Nimble}
\usage{
nimble_constants(df, interval, data_repo, post_round)
}
\arguments{
\item{df}{A data frame containing the survey data.}

\item{interval}{The length of the primary period in days.}

\item{data_repo}{The path to the data repository.}

\item{post_round}{"first" or "last"; determines which set of prior hyperparameters to use.}
}
\value{
A list of constants for the nimble model.
}
\description{
Create A List For Nimble
}
