% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble_data.R
\name{nimble_data}
\alias{nimble_data}
\title{Prepare data for nimble model}
\usage{
nimble_data(df)
}
\arguments{
\item{df}{A data frame containing the survey data.}
}
\value{
A list of data for the nimble model.
}
\description{
Prepare data for nimble model
}
