% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble_removal_model.R
\name{nimble_removal_model}
\alias{nimble_removal_model}
\title{The Nimble code for the removal model}
\usage{
nimble_removal_model()
}
\value{
model code
}
\description{
This model is based on the removal model described in Foster et al. 2025 (doi.org/10.1101/2025.10.16.680799)
}
