% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_options.R
\name{set_boaR_options}
\alias{set_boaR_options}
\title{Set options for boaR package}
\usage{
set_boaR_options(pbStyle = NULL)
}
\arguments{
\item{pbStyle}{An integer from 1 to 3 specifying the progress bar style.
Default is NULL, which will use style 3.
See ?utils::txtProgressBar for details.
When running on an HPC cluster, it is recommended to set pbStyle to 1 for better output.}
}
\description{
Set options for boaR package
}
\examples{
# Set the progress bar style to 1 for the current session
# set_boaR_options(pbStyle = 1)
}
