% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt5.R
\name{filt5}
\alias{filt5}
\title{filt5}
\usage{
filt5(lon, lat, ingrid, nodata = NA)
}
\arguments{
\item{lon}{longitude (vector) of the satellite image}

\item{lat}{latitude (vector) of the satellite image}

\item{ingrid}{The satellite data (matrix)}

\item{nodata}{value representing 'no data'}
}
\value{
returns a grid of zeros and ones
}
\description{
Finds maximum of a 5x5 sliding window. If the central pixel is the maximum, this is flagged as a one. All other pixels are flagged as zero. Usually this function is not used directly.
}
\examples{
# none
}
\seealso{
\code{\link{boa}}
}
