% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_diagnostics.R
\name{mcmc_diagnostics}
\alias{mcmc_diagnostics}
\title{Calculate MCMC diagnostics for a nimble model
Currently runs by collating chunks of MCMC
samples and then calculating diagnostics on the collated samples.}
\usage{
mcmc_diagnostics(
  mcmc_dir,
  dest,
  data,
  params_check,
  n_mcmc = 5000,
  effective_size = 1000,
  max_psrf = 15,
  verbose = TRUE,
  make_traceplot = TRUE
)
}
\arguments{
\item{mcmc_dir}{directory where MCMC chunks are stored}

\item{dest}{directory where diagnostics and posterior samples will be saved}

\item{data}{input data frame used for nimble fit}

\item{params_check}{vector of parameters (nodes) to assess convergence}

\item{n_mcmc}{number random posterior samples to save after burnin}

\item{effective_size}{minimum effective sample size for each parameter}

\item{max_psrf}{maximum potential scale reduction factor (PSRF) to assess if mcmc is behaving well.}

\item{verbose}{print diagnostic messages to the console}

\item{make_traceplot}{make traceplots for the parameters in params_check}
}
\description{
Calculate MCMC diagnostics for a nimble model
}
