% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble_inits.R
\name{nimble_inits}
\alias{nimble_inits}
\title{Prepare initial values for nimble model}
\usage{
nimble_inits(constants_nimble, data_nimble, buffer = 1000)
}
\arguments{
\item{constants_nimble}{A list of constants for the nimble model.}

\item{data_nimble}{A list of data for the nimble model.}

\item{buffer}{A numeric value to add to the initial values of N and lambda_1 to ensure they comply with take values.}
}
\value{
A list of initial values for the nimble model. Works for one chain.
}
\description{
Prepare initial values for nimble model
}
