% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_single.R
\name{single_mcmc_chain}
\alias{single_mcmc_chain}
\title{Run a single MCMC chain for the model}
\usage{
single_mcmc_chain(
  model_constants,
  model_data,
  model_flags,
  init,
  n_iter,
  chain_id = NULL,
  params_check = NULL,
  custom_samplers = NULL,
  monitors_add = NULL
)
}
\arguments{
\item{model_constants}{A list of constants for the nimble model.}

\item{model_data}{A list of data for the nimble model.}

\item{init}{A list of initial values for the nimble model.}

\item{n_iter}{The number of MCMC iterations to run.}

\item{chain_id}{Optional chain identifier used in progress output.}

\item{params_check}{Optional vector of monitored parameter nodes to print after the first sampling chunk.}

\item{custom_samplers}{A data frame specifying custom samplers to use. Should have columns "node", "type", and "control".}

\item{monitors_add}{A character vector of additional nodes to monitor.}
}
\value{
A matrix of MCMC samples.
}
\description{
Run a single MCMC chain for the model
}
